// src/app/pages/evenementiels/data-access/api.enum.ts
export const enum EndPointUrl {
    EVENEMENTIELS_CATEGORIES_ALL = "evenementiels/categories/all",
    EVENEMENTIELS_CATEGORIES_CREATE = "evenementiels/categories/create",
    EVENEMENTIELS_CATEGORIES_UPDATE = "evenementiels/categories/update",
    EVENEMENTIELS_ALL = "evenementiels/all",
    EVENEMENTIELS_ADD = "evenementiels/add",
    EVENEMENTIELS_UPDATE = "evenementiels/update",
    EVENEMENTIELS_DELETE = "evenementiels/delete",
    EVENEMENTIELS_ROLES_ALL = "evenementiels/roles/all",
    ARTISTS_CATEGORIES_ALL = "artists/categories/all",
}